% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Solve ODE for simple odem model}
\usage{
run_model(bc, params, ini, times)
}
\arguments{
\item{bc}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{params}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{ini}{matrix; Water temperatures (rows correspond to time, cols to depth)}

\item{times}{vector having time information}
}
\value{
results for oxgen concentrations
}
\description{
Extracts time (from date column) and space (aka depth) information
}
